+++ runtime/core/inference_engine/unit-tests/ie_blob_test.cpp
@@ -338,7 +338,7 @@ TEST_F(BlobTests, canModifyDataInRangedFor) {
         data = 5;
     }
 
-    for (int i = 0; i < v.size(); i++) {
+    for (size_t i = 0; i < v.size(); i++) {
         ASSERT_EQ(5, blob.data()[i]) << "Mismatch at" << i;
     }
 }
