+++ tests/tests_deprecated/functional/vpu/common/layers/myriad_layers_scatter_update_test.hpp
@@ -280,7 +280,7 @@ private:
     std::string to_string(const SizeVector& v) {
         std::stringstream s;
         s << "{";
-        for (int i = 0; i < v.size(); i++) {
+        for (size_t i = 0; i < v.size(); i++) {
             s << (i? ", ": "") << v[i];
         }
         s << "}";
